coure formula + some fixes

+++ AcademicManagementBackEnd/Entities/FinalGrade.cs
@@ -6,6 +6,6 @@ namespace Entities
     {
         public Guid CourseId { get; set; }
         public Guid StudentId { get; set; }
-        public float Value { get; set; }
+        public double Value { get; set; }
     }
 }
